AMQP protocol permits bodyless messages
"Certain methods (such as Basic.Publish, Basic.Deliver, etc.) are
formally defined as carrying content. When a peer sends such a method
frame, it always follows it with a content header and *zero* or more
content body frames."

So don't push body frame parser immediately, but only when header says
that body size is greater that zero.