old check was triggering only if on_return is provided
mixing '!/&&/||' and 'or/and/not' in one condition always bad
idea. Mixing 'and' with 'or' without parens even worse. Old
code was triggering error in one case as it works like this:

    unless ( $confirm ) {
        $ack or ( $nack or ( $return and $croak ) )
    }

So if we are not in confirm mode then croak if and only if return
callback is set, but not the others.

Use two conditions with nested block instead. I find it more readable
as well as correct.